backend/ecommerce/productos/views.py
from django.shortcuts import render
from rest_framework import generics
from .serializers import ProductosSerializer
from .models import Productos

class ProductosList(generics.ListCreateAPIView):
    queryset = Productos.objects.all()
    serializer_class = ProductosSerializer

class ProductosDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Productos.objects.all()
    serializer_class = ProductosSerializer
