backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.1
sqlparse==0.5.0
tzdata==2024.1
Pillow==10.4.0 #para el manejo de imagenes